Built the website and dragged in some project
I built the website.
Simple design with a header and paragraph beneath.
Nav-bar made with grid and lastly an image at the bottom.

Links to 2 projects:
JS Roulette: A small project where I refreshes my JS knowledge and built a simple roulette-style game.

U1: My first assignment for the DDU1 course - a little introduction about myself and my hobbies.

Lastly i built a status-page for the "Coming soon" links that just is a header and paragraph and nav-bar in the same style as the homepage.